+++ src/sprtf.cpp
@@ -23,6 +23,9 @@
  *   along with this program; if not, write to the Free Software
  *   Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, US
  *
+ *
+ * Modified 2016-09-09 by Bob Rudis for the R package htmltidy
+ *
  */
 
 #ifdef _MSC_VER